Cambios
Se realiza corrección en clase Logic de dos clases
Se crean todas las clases Resource Faltantes

+++ src/main/java/edu/konrad/sistemacalificaciones/logic/PersonaLogic.java
@@ -66,7 +66,7 @@ public class PersonaLogic {
      * @param personaActualizar
      * @return persona actualizado
      */
-    public PersonaEntity actualizarPersona(Long id, PersonaEntity personaActualizar) {
+    public PersonaEntity actualizarPersona(int id, PersonaEntity personaActualizar) {
         PersonaEntity personaUpdate = personaPersistence.update(personaActualizar);
         return personaUpdate;
     }
